Move and document Webfinger more
