+++ src/Api/Site/Webfinger.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace Smolblog\Api\User;
+namespace Smolblog\Api\Site;
 
 use Smolblog\Api\Endpoint;
 use Smolblog\Api\EndpointConfig;
-use Smolblog\Api\GenericResponse;
 use Smolblog\Api\ParameterType;
 use Smolblog\Framework\Objects\Identifier;
-use Smolblog\Framework\Objects\Value;
 
 /**
  * Find a user with Webfinger
  *
  * Implementation of the [Webfinger] standard. This protocol is used by Mastodon to enable user lookups from its
- * `@user@server` mention format.
+ * `@user@server` mention format. Smolblog considers interactions to originate from a site, so this will attempt to
+ * find the site described.
  *
  * [Webfinger]: https://webfinger.net
  */
@@ -30,19 +29,19 @@ class Webfinger implements Endpoint {
 				'resource' => ParameterType::required(ParameterType::string()),
 				'rel' => ParameterType::string(),
 			],
-			responseShape: ParameterType::object(
-				subject: ParameterType::string(),
-				links: ParameterType::array(ParameterType::object(
-					rel: ParameterType::string(),
-					type: ParameterType::string(),
-					href: ParameterType::string(),
-				))
-			),
 			requiredScopes: [],
 		);
 	}
 
-	public function run(?Identifier $userId, ?array $params, ?object $body): WebfingerResponse {
+	/**
+	 * Run the endpoint.
+	 *
+	 * @param Identifier|null $userId Ignored.
+	 * @param array|null      $params Expects 'resource'.
+	 * @param object|null     $body   Ignored.
+	 * @return WebfingerResponse
+	 */
+	public function run(?Identifier $userId = null, ?array $params, ?object $body = null): WebfingerResponse {
 		return new WebfingerResponse(
 			subject: $params['resource'],
 		);
